# swagger.yaml
openapi: 3.0.0
info:
  title: Personal Finance Tracker API
  version: 1.0.0
  description: REST API for managing personal finance transactions, including income and expenses.
servers:
  - url: http://localhost:3000
paths:
  /transactions: 
    get:
      summary: Retrieve a list of all transactions with pagination support.
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: The page number to retrieve.
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
          description: The number of transactions to retrieve per page.
      responses: 
        '200':
          description: Successfully retrieved the list of transactions.
    post:
      summary: Create a new transaction (income or expense).
      requestBody:
        description: The transaction to create.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: integer
                  description: The amount of the transaction.
                type:
                  type: string
                  enum: [income, expense]
                  description: The type of the transaction (income or expense).
                category:
                  type: string
                  description: The category of the transaction (e.g., groceries, salary).
                date:
                  type: string
                  format: date-time
                  description: The date and time of the transaction.
      responses: 
        '201':
          description: Successfully created a new transaction.

  /categories:
    get:
      summary: Retrieve a list of all categories.
      responses:
        '200':
          description: Successfully retrieved the list of categories.
    post:
      summary: Create a new category.
      requestBody:
        description: The category to create.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The name of the category (e.g., Groceries, Salary, Rent).
      responses:
        '201':
          description: Successfully created a new category.

  /categories/{id}:
    get:
      summary: Retrieve a single category by ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the category to retrieve.
      responses:
        '200':
          description: Successfully retrieved the category.
        '404':
          description: Category not found.
    put:
      summary: Update a single category by ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the category to update.
      requestBody:
        description: The updated category data.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The updated name of the category.
      responses:
        '200':
          description: Successfully updated the category.
        '404':
          description: Category not found.
    delete:
      summary: Delete a single category by ID.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The ID of the category to delete.
      responses:
        '200':
          description: Successfully deleted the category.
        '404':
          description: Category not found.